+++ integration/stubs/assets.go
@@ -13,6 +13,7 @@
 package stubs
 
 import (
+	"context"
 	"errors"
 
 	"code.vegaprotocol.io/vega/assets"
@@ -62,19 +63,21 @@ func (a *AssetStub) SetStrict() {
 	a.permissive = false
 }
 
-func (AssetStub) Enable(assetID string) error {
+func (AssetStub) Enable(_ context.Context, assetID string) error {
 	return nil
 }
 
 type isAssetStub struct {
 	ID            string
 	DecimalPlaces uint64
+	Status        types.AssetStatus
 }
 
 func NewIsAssetStub(id string, dp uint64) *assets.Asset {
 	return assets.NewAsset(&isAssetStub{
 		ID:            id,
 		DecimalPlaces: dp,
+		Status:        types.AssetStatusProposed,
 	})
 }
 
@@ -86,6 +89,18 @@ func (a isAssetStub) Type() *types.Asset {
 			Quantum:  num.DecimalFromFloat(5000),
 		},
 	}
+}
+
+func (a *isAssetStub) SetPendingListing() {
+	a.Status = types.AssetStatusPendingListing
+}
+
+func (a *isAssetStub) SetRejected() {
+	a.Status = types.AssetStatusRejected
+}
+
+func (a *isAssetStub) SetEnabled() {
+	a.Status = types.AssetStatusEnabled
 }
 
 func (isAssetStub) GetAssetClass() common.AssetClass {