+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - [5508](https://github.com/vegaprotocol/vega/issues/5508) - Fix duplicated staking events
 - [5514](https://github.com/vegaprotocol/vega/issues/5514) - Emit `rewardScore` event correctly after loading from checkpoint
 - [5520](https://github.com/vegaprotocol/vega/issues/5520) - Do not fail silently when wallet fails to start
+- [5521](https://github.com/vegaprotocol/vega/issues/5521) - Fix asset bundle and add asset status
 
 ## 0.52.0